requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
asyncmy==0.2.9
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.2
fastapi-cli==0.0.5
fastapi-versioning==0.10.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
fonttools==4.54.1
fpdf2==2.8.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.21.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.28.0
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
lxml_html_clean==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.10
mypy==1.16.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.1
packaging==24.1
parsel==1.9.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
Protego==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.14.3
queuelib==1.7.0
redis==6.2.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.2
rsa==4.9
ruff==0.12.2
Scrapy==2.11.2
scrapy-splash==0.9.0
service-identity==24.1.0
setuptools==75.5.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.40.0
tenacity==8.5.0
tinysegmenter==0.3
tldextract==5.1.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.7.0
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
validators==0.34.0
w3lib==2.2.1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
wheel==0.45.0
zope.interface==7.1.0
